Update `README.md` (#6)
# Pull Request

## Description

This pull request makes minor updates to the project documentation and
configuration files. The changes include renaming the repository title
in the `README.md` file and updating the minimum version requirement for
`lefthook` in its configuration file.

Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L1-R5):
Changed the repository title from "Repository Name" to "Projects" and
updated the corresponding table of contents entry.

Configuration updates:

*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the `min_version` of `lefthook` from `1.11.12` to `1.11.13`.