Add First Status (#5)
# Pull Request

## Description

This pull request adds a new `STATUSES.md` file to document the status
of projects. The most important change is the addition of a
"Sunsetted/Decommissioned" section to indicate that the project is no
longer maintained.

Documentation updates:

*
[`STATUSES.md`](diffhunk://#diff-d7e8611a894fc6c22d6f911179e51f1705a728a93b5f7993016285acb7b3d003R1-R8):
Created a new file with a "Statuses" table of contents and a
"Sunsetted/Decommissioned" section, including a warning that the project
is no longer maintained.