Add link to statuses (#11)
# Pull Request

## Description

This pull request updates the `README.md` file to include a new
"Statuses" section, which provides information about project statuses
and links to the `STATUSES.md` file for further details.

Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R7-R12):
Added a "Statuses" section to describe the current state of a project
and included a link to the `STATUSES.md` file.