Add New Status (#19)
# Pull Request

## Description

This pull request updates the `STATUSES.md` file to introduce a new
project status called "Terminated." The addition clarifies the state of
projects that are halted mid-development.

Key change:

* Added a new "Terminated" status to the project statuses table to
indicate projects that are stopped mid-development. This includes a
badge and description specifying that the repository will be archived if
not already. (`STATUSES.md`,
[STATUSES.mdR15](diffhunk://#diff-d7e8611a894fc6c22d6f911179e51f1705a728a93b5f7993016285acb7b3d003R15))