Update Reusable Workflows (#28)
# Pull Request

## Description

This pull request updates the reusable workflow references across
multiple GitHub Actions workflow files to use a newer version
(`v2025.05.24.02`) of the workflows. This ensures that the latest
improvements and fixes in the reusable workflows are applied.

### Workflow Reference Updates:

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dL15-R15):
Updated the reusable workflow reference for `common-clean-caches.yml` to
`v2025.05.24.02`.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L42-R42):
Updated the reusable workflow references for both
`common-code-checks.yml` and `codeql-analysis.yml` to `v2025.05.24.02`.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L42-R42)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L54-R54)
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL15-R15):
Updated the reusable workflow reference for
`common-pull-request-tasks.yml` to `v2025.05.24.02`.
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L18-R18):
Updated the reusable workflow reference for `common-sync-labels.yml` to
`v2025.05.24.02`.