Add development-environment project status (#55)
# Pull Request

## Description

This pull request updates the `README.md` file to improve the clarity
and consistency of project status indicators. The most important change
is the addition of a status badge for the `tech-radar` project.

Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L49-R49):
Added a "Maintenance" status badge to the `tech-radar` project entry to
align with other projects and provide consistent status information.